press run button to call backend
You can now press the run button to run program.I also fixed the save changes and close button on model because when using the run button it has to change due to user error. User would be able to use a ref photo and video that was preveiously not wanted. changes came in the Uploadmodel function and i created a new function savemodal. I aslo created a runFunction